+++ lib/state.dart
@@ -39,6 +39,7 @@ class GlobalState {
   late CommonTheme theme;
   late Color accentColor;
   bool needInitStatus = true;
+  // ignore: deprecated_member_use
   CorePalette? corePalette;
   DateTime? startTime;
   UpdateTasks tasks = [];